Added a propper check for a sleep check.
It turns out this is a bug with the client, and spigot had the same
problem. Spigot admins ended up just decreasing the max bulk chunk
size to fix it. I do not have a better fix other than to just slow
down the data going to the client.

It is appalling to me that this got into production code, but I
can't exactly patch the 1.8 and 1.12 minecraft clients, so I guess
this is going to be a permanent change. I have the 1.8 client
decompiled, so I am going to try to find the bug in that (to more
correctly tune the sleep statement in the server).