The server binary compiles. Nothing works.
Some things work. Terrain generates on another thread, and we can listen
for clients. However, the only reason that works is because we haven't
run out of tokio threads. That terrain generation is a blocking thread
running in a tokio task. Needless to say, this needs a lot of fixing.