Replaced all the packets in the server with the new generalized standard
This will help a lot with reducing complexity in the server. Every
packet now has one version, and the proxy will manage all the version
seperation. This is more code in total, and will probably be more work,
but it lets the server be much simpler in the end.

The proxy will need to store all the block cross-versioning data. This
means that the server only needs to work with latest block ids, which
will simplify things a lot. However, this also means the proxy will need
to be updated when a new minecraft version is released. I don't really
see a way around this, and I'm just going to live with it. They proxy
would need to be updated for new packets anyways, so it won't make too
much difference.